ajout de tous les apis familly

+++ heritagehub/heritagehubapp/views/PersonViewSet.py
@@ -14,7 +14,7 @@ class PersonViewSet(viewsets.ModelViewSet):
     permission_classes=[IsAuthenticated]
 
     def get_permissions(self):
-        if self.action =='list':
+        if self.action =='list' or self.action == 'retrieve':
             permission_class = [IsGetRequest]
         else:
             permission_class = [IsAuthenticated]
@@ -42,17 +42,23 @@ class PersonViewSet(viewsets.ModelViewSet):
       return Response(serializer.data, status=status.HTTP_200_OK)
 
     def partial_update(self, request, *args, **kwargs):
-         person_id = kwargs['pk']
-         person = get_object_or_404(PersonModel, pk=person_id)
+         if self.request.user.id == request.data['created_by']:
+            person_id = kwargs['pk']
+            person = get_object_or_404(PersonModel, pk=person_id)
 
-         serializer = self.get_serializer(person, data=request.data, partial=True)
-         serializer.is_valid(raise_exception=True)
-         self.perform_update(serializer)
-         return Response(serializer.data, status=status.HTTP_200_OK)
+            serializer = self.get_serializer(person, data=request.data, partial=True)
+            serializer.is_valid(raise_exception=True)
+            self.perform_update(serializer)
+            return Response(serializer.data, status=status.HTTP_200_OK)
+         else:
+            return Response({"message":"permission denied"}, status=status.HTTP_401_UNAUTHORIZED)
 
     def destroy(self, request, *args, **kwargs):
-      person_id = kwargs['pk']
-      person = get_object_or_404(PersonModel, pk=person_id)
+      if self.request.user.id == request.data['created_by']:
+        person_id = kwargs['pk']
+        person = get_object_or_404(PersonModel, pk=person_id)
 
-      person.delete()
-      return Response( status=status.HTTP_200_OK)
+        person.delete()
+        return Response( status=status.HTTP_200_OK)
+      else:
+            return Response({"message":"permission denied"}, status=status.HTTP_401_UNAUTHORIZED)
